+++ CHANGELOG.md
@@ -7,15 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-## Changed
+### Changed
 
 - Remove `<MajorityVotingResult />` and `<ApprovalThresholdResult />` sub-components for `<ProposalDataListItem />`
 - Update `InputFileAvatar` core component to remove `onFileSelect` property and add `value` and `onChange` properties
 - Update `IProposalActionUpdateMetadataDaoMetadata` interface logo property to `avatar` to better align with actions
-- Update `<ProposalDataListItem />` module component to include `statusContext` property
-- Update `<ProposalDataListItem />` module component to support `ADVANCEABLE` status
 - Bump `softprops/action-gh-release` from 2.2.0 to 2.2.1
 - Update minor and patch NPM dependencies
+- Update `<ProposalDataListItem />` module component to include `statusContext` property
+- Update `<ProposalDataListItem />` module component to support `ADVANCEABLE` status
 
 ## [1.0.62] - 2025-01-08